requirements.txt
apache-airflow==2.9.2
apache-airflow-providers-postgres==2.0.1
psycopg2-binary==2.9.5
requests==2.28.2
python-dotenv==0.21.1
